make design

+++ db/schema.rb
@@ -2,8 +2,8 @@
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
 #
-# This file is the source Rails uses to define your schema when running `rails
-# db:schema:load`. When creating a new database, `rails db:schema:load` tends to
+# This file is the source Rails uses to define your schema when running `bin/rails
+# db:schema:load`. When creating a new database, `bin/rails db:schema:load` tends to
 # be faster and is potentially less error prone than running all of your
 # migrations from scratch. Old migrations may fail to apply correctly if those
 # migrations use external dependencies or application code.
@@ -15,8 +15,8 @@ ActiveRecord::Schema.define(version: 2021_03_19_020355) do
   create_table "comments", force: :cascade do |t|
     t.string "content"
     t.integer "recommend"
-    t.integer "user_id", null: false
-    t.integer "place_id", null: false
+    t.bigint "user_id", null: false
+    t.bigint "place_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["place_id"], name: "index_comments_on_place_id"
@@ -36,7 +36,7 @@ ActiveRecord::Schema.define(version: 2021_03_19_020355) do
   create_table "places", force: :cascade do |t|
     t.string "title"
     t.text "content"
-    t.integer "user_id", null: false
+    t.bigint "user_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.string "image"
@@ -62,8 +62,8 @@ ActiveRecord::Schema.define(version: 2021_03_19_020355) do
   end
 
   create_table "tag_maps", force: :cascade do |t|
-    t.integer "place_id", null: false
-    t.integer "tag_id", null: false
+    t.bigint "place_id", null: false
+    t.bigint "tag_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["place_id"], name: "index_tag_maps_on_place_id"
